move all of search to search_dropdown, and make it work
